+++ zeroCore/src/File/Read/PC/ZC_PC_FileReader.h
@@ -21,7 +21,7 @@ public:
 
 private:
     std::ifstream file;
-    const size_t size;
+    const long size;
 
     bool OpenCheck() const;
     long CalculateSize();
